requirements.txt
requests
openai
dotenv
pandas
schedule
snowflake-connector-python[pandas]